backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
pymongo==4.6.0
dnspython>=2.0.0
Pillow==10.1.0
opencv-python==4.8.1.78
PyMuPDF==1.23.8
python-docx==1.1.0
pytesseract==0.3.10
easyocr==1.7.0
numpy==1.24.3
cryptography==41.0.7
pandas==2.1.3
openpyxl==3.1.2
reportlab==4.0.7
Werkzeug==3.0.1
bcrypt==4.1.2
certifi>=2023.0.0
twilio==8.10.0
PyJWT==2.8.0
redis==5.0.1
psutil==5.9.6
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
PyPDF2==3.0.1
pdfplumber==0.9.0
cachetools==5.3.2
razorpay==1.4.2

